worker/requirements.txt
redis==5.0.0
psycopg2-binary==2.9.7
ffmpeg-python==0.2.0
openai-whisper==20231117
torch --index-url https://download.pytorch.org/whl/cpu
numpy
